refactor: integration test suite
- Move helper functions to own file
- Use physical rather than logical CPU cores for parallel runs
- Add verbose flag when running on CI, which makes it much easier to debug failures
