+++ pixi.toml
@@ -39,12 +39,12 @@ rich = ">=13.7.1,<14"
 tomli-w = ">=1.0,<2"
 
 [feature.pytest.tasks]
-test-common-wheels-ci = { cmd = "pytest -n logical tests/wheel_tests/" }
-test-common-wheels-dev = { cmd = "pytest -n logical tests/wheel_tests/", depends-on = [
+test-common-wheels-ci = { cmd = "pytest --numprocesses=auto --verbose tests/wheel_tests/" }
+test-common-wheels-dev = { cmd = "pytest --numprocesses=auto tests/wheel_tests/", depends-on = [
   "build",
 ] }
-test-integration-ci = "pytest -n logical tests/integration"
-test-integration-dev = { cmd = "pytest -n logical tests/integration", depends-on = [
+test-integration-ci = "pytest --numprocesses=auto --verbose tests/integration"
+test-integration-dev = { cmd = "pytest --numprocesses=auto tests/integration", depends-on = [
   "build",
 ] }
 typecheck-integration = "mypy --strict tests/integration"
